Verify examples
I've been thinking about how to optimize this a bit. I converted the easy parts over to pure numpy, and then verified that the examples are still able to train correct. The next step is to convert the backprop into something numpy can handle efficiently, but backpropagating over an arbitrary graph isn't quite as nice as backpropagating down layers.