Fix and optimize backprop
I reworked the way I set up the graph so that the order of rows and columns were arranged like a Markov chain transition matrix. I've been working on a more compact numpy-friendly way of doing backpropagation on an arbitrary graph, and I think what I have here is pretty close, but there may still be some issues to work out. Along the way, I happened to find what appears to be a bug in numpy...